# openspec/changes/add-adaptive-model-router/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Coordinator Model Routing API
  version: 1.0.0
  description: >
    Adaptive model selection, catalog, decisions, usage, and feedback surface.
    Bearer-authenticated like all coordinator endpoints. MCP tool
    `select_model_for_task` mirrors POST /routing/select_model.
security:
  - CoordinatorBearer: []
paths:
  /routing/select_model:
    post:
      operationId: selectModelForTask
      summary: Rank feasible candidates and select a model for a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectModelRequest'
      responses:
        '200':
          description: Selection with provenance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelectModelResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/RoutingUnavailable'
        '422':
          $ref: '#/components/responses/Problem'
  /routing/catalog:
    get:
      operationId: listCatalog
      summary: List catalog rows (no external calls on read path)
      parameters:
        - name: endpoint_kind
          in: query
          schema: {$ref: '#/components/schemas/EndpointKind'}
        - name: available_only
          in: query
          schema: {type: boolean, default: true}
      responses:
        '200':
          description: Catalog rows
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/CatalogRow'}
        '401':
          $ref: '#/components/responses/Unauthorized'
  /routing/decisions/{decision_id}:
    get:
      operationId: getRoutingDecision
      summary: Fetch a decision provenance record
      parameters:
        - name: decision_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Provenance record
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RoutingDecision'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/Problem'
  /routing/usage:
    get:
      operationId: getRoutingUsage
      summary: Aggregated spend, counterfactual savings, exploration burn-down
      parameters:
        - name: window
          in: query
          schema: {type: string, enum: [day, week, month], default: month}
        - name: include_estimates
          in: query
          schema: {type: boolean, default: true}
      responses:
        '200':
          description: Usage aggregates
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UsageAggregate'}
        '401':
          $ref: '#/components/responses/Unauthorized'
  /routing/feedback:
    post:
      operationId: postRoutingFeedback
      summary: Accept best-effort task-outcome feedback for audit
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FeedbackEvent'}
      responses:
        '202':
          description: Accepted for best-effort audit recording
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    CoordinatorBearer:
      type: http
      scheme: bearer
      bearerFormat: API key
  responses:
    Problem:
      description: RFC 7807 problem
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type: {type: string}
              title: {type: string}
              status: {type: integer}
              detail: {type: string}
    Unauthorized:
      description: Missing or invalid coordinator API key
      content:
        application/json:
          schema: {$ref: '#/components/schemas/HTTPError'}
    RoutingUnavailable:
      description: No feasible model-routing candidate is available
      content:
        application/json:
          schema: {$ref: '#/components/schemas/HTTPError'}
  schemas:
    HTTPError:
      type: object
      required: [detail]
      properties:
        detail: {type: string}
    EndpointKind:
      type: string
      enum: [vendor-cli, vendor-sdk, openrouter, local]
    ObjectiveProfile:
      type: string
      enum: [quality-first, balanced, cost-first, resilience]
    SelectModelRequest:
      type: object
      required: [task_signals]
      properties:
        task_signals:
          type: object
          required: [archetype]
          properties:
            archetype: {type: string, example: implementer}
            phase: {type: [string, "null"], default: null, example: IMPLEMENT}
            task_type: {type: [string, "null"], default: null, example: implementer/high-complexity}
            complexity:
              type: [string, "null"]
              enum: [low, medium, high, null]
              default: null
            modality: {type: string, enum: [interactive, programmatic], default: programmatic}
          additionalProperties: true
        objective_profile:
          anyOf:
            - {$ref: '#/components/schemas/ObjectiveProfile'}
            - {type: "null"}
          default: null
        weight_overrides:
          anyOf:
            - type: object
              properties:
                w_quality: {type: [number, "null"], default: null}
                w_cost: {type: [number, "null"], default: null}
                w_latency: {type: [number, "null"], default: null}
            - {type: "null"}
          default: null
        allow_exploration: {type: boolean, default: true}
      example:
        task_signals: {archetype: implementer, phase: IMPLEMENT, complexity: low, modality: programmatic}
        objective_profile: cost-first
    Candidate:
      type: object
      required: [vendor, model, endpoint_kind, score]
      properties:
        vendor: {type: string, example: openrouter}
        model: {type: string, example: qwen/qwen3-coder}
        endpoint_kind: {$ref: '#/components/schemas/EndpointKind'}
        score: {type: number, example: 0.71}
        quality: {type: [number, "null"], default: null}
        norm_cost: {type: [number, "null"], default: null}
        norm_latency: {type: [number, "null"], default: null}
        posterior_sample_size:
          type: [number, "null"]
          default: null
          description: Effective sample mass after source weighting and time decay; may be fractional.
        stale_catalog: {type: boolean, default: false}
        cost_source:
          type: [string, "null"]
          enum: [posterior, prior, null]
          default: null
          description: >
            Provenance of the cost estimate used in scoring — `posterior` (learned
            per-task cost cleared the sample threshold) or `prior` (catalog per-Mtok
            proxy). Null when no cost was scored. Mirrors the resolver Candidate.
    SelectModelResponse:
      type: object
      required: [decision_id, selected, alternatives]
      properties:
        decision_id: {type: string, format: uuid}
        selected: {$ref: '#/components/schemas/Candidate'}
        alternatives:
          type: array
          items: {$ref: '#/components/schemas/Candidate'}
        exploration: {type: boolean, default: false}
        fallback:
          type: boolean
          default: false
          description: True when static-tier resolution was returned (flag off / error / timeout)
        excluded:
          type: array
          default: []
          description: Candidates removed pre-scoring by hard constraints, with reasons
          items:
            type: object
            required: [vendor, model, reason]
            properties:
              vendor: {type: string}
              model: {type: string}
              reason: {type: string, example: cedar:programmatic-ineligible}
      example:
        decision_id: 3f1b7c1e-9a3d-4f9d-8b1a-2c6e5d4a9b0c
        selected: {vendor: local, model: qwen3-coder-32b, endpoint_kind: local, score: 0.66}
        alternatives: []
        exploration: true
        fallback: false
        excluded: [{vendor: claude, model: fable-5, reason: cedar:programmatic-ineligible}]
    CatalogRow:
      type: object
      required: [vendor, model, endpoint_kind, available]
      properties:
        vendor: {type: string}
        model: {type: string}
        endpoint_kind: {$ref: '#/components/schemas/EndpointKind'}
        base_url: {type: [string, "null"], default: null, example: http://localhost:11434/v1}
        prompt_usd_per_mtok: {type: [number, "null"], default: null}
        completion_usd_per_mtok: {type: [number, "null"], default: null}
        context_window: {type: [integer, "null"], default: null}
        benchmark_priors:
          type: object
          default: {}
          additionalProperties: {type: number}
          example: {artificial_analysis_intelligence: 62.5}
        p50_latency_ms: {type: [number, "null"], default: null}
        available: {type: boolean}
        quota_headroom_pct: {type: [number, "null"], default: null, description: Proactive remaining-window pct (design D13); null when quota probe off/uncovered}
        quota_reset_at: {type: [string, "null"], format: date-time, default: null}
        quota_source: {type: [string, "null"], default: null, example: quota-axi}
        refreshed_at: {type: [string, "null"], format: date-time, default: null}
        stale: {type: boolean, default: false}
    RoutingDecision:
      type: object
      required: [decision_id, created_at, request, selected, policy_version]
      properties:
        decision_id: {type: string, format: uuid}
        created_at: {type: string, format: date-time}
        request: {$ref: '#/components/schemas/SelectModelRequest'}
        selected: {$ref: '#/components/schemas/Candidate'}
        alternatives:
          type: array
          default: []
          items: {$ref: '#/components/schemas/Candidate'}
        exploration: {type: boolean, default: false}
        fallback: {type: boolean, default: false}
        policy_version: {type: string}
        excluded:
          type: array
          default: []
          items:
            type: object
            required: [vendor, model, reason]
            properties:
              vendor: {type: string}
              model: {type: string}
              reason: {type: string}
        budget_state:
          type: object
          default: {}
          properties:
            exploration_pct_used: {type: [number, "null"], default: null}
            exploration_usd_used: {type: [number, "null"], default: null}
            metered_usd_used: {type: [number, "null"], default: null}
        outcome_ref:
          type: [string, "null"]
          default: null
          description: Link to realised outcome (feedback event id) once known
    UsageAggregate:
      type: object
      required: [window]
      properties:
        window: {type: string}
        by_model:
          type: array
          default: []
          items:
            type: object
            required: [vendor, model, endpoint_kind]
            properties:
              vendor: {type: string}
              model: {type: string}
              endpoint_kind: {$ref: '#/components/schemas/EndpointKind'}
              prompt_tokens: {type: integer, default: 0}
              completion_tokens: {type: integer, default: 0}
              actual_usd: {type: number, default: 0}
              counterfactual_usd: {type: number, default: 0}
              estimated_fraction: {type: number, default: 0, description: Share of spend from estimated token counts}
        net_savings_usd: {type: number, default: 0}
        net_savings_usd_excluding_estimates: {type: number, default: 0}
        exploration_usd_used: {type: number, default: 0}
        metered_ceiling_usd: {type: number, default: 0}
        metered_usd_used: {type: number, default: 0}
        estimated_fraction: {type: number, default: 0, description: Share of included entries with estimated token counts}
    FeedbackEvent:
      type: object
      description: >-
        The dg-00 endpoint accepts and audits this envelope best-effort. It does not
        associate feedback with a catalog row or update posteriors; vendor and model
        are the observed dispatch identity, and optional decision_id supplies
        provenance when available. Endpoint-disambiguated aggregation is deferred.
      required: [source, vendor, model, task_type, metrics]
      properties:
        decision_id: {type: [string, "null"], format: uuid, default: null}
        source:
          type: string
          enum: [gen-eval, validation, vendor-switch, procedural-memory, transcript-triage]
        vendor: {type: string}
        model: {type: string}
        task_type: {type: string}
        metrics:
          type: object
          properties:
            success: {type: [boolean, "null"], default: null}
            quality_score: {type: [number, "null"], default: null}
            cost_observed_usd: {type: [number, "null"], default: null}
            latency_observed_seconds: {type: [number, "null"], default: null}
            tokens_estimated: {type: boolean, default: false}
        observed_at: {type: [string, "null"], format: date-time, default: null}